+++ wp-content/themes/snc-blog-child/functions.php
@@ -3,10 +3,10 @@
 add_action( 'wp_enqueue_scripts', 'snc_blog_child_enqueue_styles' );
 
 function snc_blog_child_enqueue_styles() {
-  wp_enqueue_style( 'parent-style', get_template_directory_uri() . '/style.css' );
-  wp_enqueue_style( 'child-style',
+  wp_enqueue_style( 'snc-blog-child',
     get_stylesheet_directory_uri() . '/style.css',
-    array('parent-style')
+    array('snc-blog'),
+    wp_get_theme()->get( 'Version' )
   );
 }
 
@@ -27,7 +27,8 @@ add_action('admin_head', 'custom_dashboard_css');
 function custom_dashboard_css() {
   // Hide all plugin registration notices after installing WP-Optimize Premium
   echo '<style>
-    .updated#udmupdater_not_connected {
+    .updated#udmupdater_not_connected,
+  .error.updraftmanagermessage {
       display: none !important;
       visibility: hidden !important;
     }